Problem: performance of generator termination
There's a penalty associated with returning from an async function.

I've learned this from @kprotty (thank you!) that if one to provide a
return value across suspension instead of returning from an async function,
"it [will] avoid rmw + full barrier which can slow down single threaded code in
the hot path + prevent compiler optimizations which act on the return value"

Solution: provide `Handle.finish()` helper to those who need extra performance

This also changes the signature of `Handle(T)` to `Handle(T, Return)` so it is
a breaking change.

Note that `Handle.finish` should be `noreturn` but it is not at the moment because
of a bug in the compiler https://github.com/ziglang/zig/issues/5728