Problem: access to return_value is complicated
One needs to go through a pointer and an optional. It's
excessive and is hard to read.

Solution: hinge return_value validity on `Generator.state`

`Generator.return_value`'s value is undefined until `Generator.state`
becomes `.Done` and no error was raised, at which point it's supposed to have a
value.