Problem: unsafe access to undefined Generator.return_value
While it has been noted in the documentation that this value
is only valid if generator successfully ended, it is still
a potential source for mistakes.

Solution: move return value into .state.Done
so it can only be accessed if the union is set to .Done

This also adds .Error tag into the union to distinguish
successful terminations from unsuccessful ones.