Problem: how slow are generators?
It's not clear what is the impact of using generators.

Solution: add a comparison benchmark for generators vs callbacks

Early, unscientific and synthetic benchmarking suggests that empty
callbacks are twice as fast as draining generators. The utility of either
isn't high as we are not doing anything with these yields. My current
hypothesis is that once there's going to be real processing, the differences
won't be as pronounced simply because they will drown in the noie of real work.

Still, it's an important observation to note. Pure overhead of generators vs
callbacks *seems* to be 100% (twice as slow). Maybe this can be improved down
the road?

Also, while at it, compile benchmarks for as `.ReleaseFast`,
otherwise we're not capturing useful measurements.