Problem: what's the point of generators?
Generators look nice, but they are still slower than callbacks
(as they naturally have more harness code).

Solution: provide Map and Join generators

The beauty of generators is that they are async and this allows us
to build higher-level primitives.

Of a particular interest is Join that will allow to consume values
from two generators, as they come.