Problem: occassional test crashes
Trying to resume a non-suspended async frame

Solution: ensure validity of frame's suspension status

If generator has yielded before frame suspension, we were not
setting the status back to no .Unsuspended. We were only doing
this if the generator was still in .Working before the suspension.

Now we're making sure the status is always .Unsuspended past the
point of need of suspension -- whether the suspension happened or not.