Problem: Map generators are purely functional
That is, it takes input and produces an output. Which is beautiful
but also limiting in terms of usefulness.

Solution: allow creation of stateful Map generators

These generators require the initial state and they'll pass
it to every invocation of the mapping function, allowing it to
maintain state and be useful.

Developed during a live coding session: https://youtu.be/9BTW4qOp0TQ