Problem: limited utility in allocating Join generator
Recent addition of `allocating` option to Join generator
might be of dubious value.

Instead of having the generator allocate each frame individually,
wouldn't it often be more beneficial to simply allocate the
entire generators? Then it'll be just one allocation and it'll
be laid out linearly.

The only benefit of this option I can see is that it won't allocate
until we start iterating over yielded values.

Solution: make it de-allocate frame memory eagerly

To make allocating Join generator more practical, let it de-allocate an
allocated frame as soon as its corresponding generator is done. There's no need
to keep that allocation around.

This increases the utility of the allocating Join generator as it'll
release memory sooner, decreasing overall memory pressure.